ServiceAwardCardをServiceCardに統一
